Reorganize TagInfo producers and JetTags producers.
