+++ RecoBTag/FeatureTools/plugins/BuildFile.xml
@@ -1,10 +1,9 @@
 <use name="FWCore/Framework"/>
 <use name="FWCore/MessageLogger"/>
-<library file="*.cc" name="RecoBTagDeepBoostedJetPlugins">
+<library file="*.cc" name="RecoBTagFeatureToolsPlugins">
   <use name="DataFormats/BTauReco"/>
   <use name="DataFormats/Common"/>
   <use name="RecoBTag/FeatureTools"/>
   <use name="DataFormats/PatCandidates"/>
-  <use name="PhysicsTools/MXNet"/>
   <flags EDM_PLUGIN="1"/>
 </library>
