+++ RecoBTag/FeatureTools/python/pfNegativeDeepFlavourTagInfos_cfi.py
@@ -1,6 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 
-from RecoBTag.TensorFlow.pfDeepFlavourTagInfos_cfi import pfDeepFlavourTagInfos
+from RecoBTag.FeatureTools.pfDeepFlavourTagInfos_cfi import pfDeepFlavourTagInfos
 
 pfNegativeDeepFlavourTagInfos = pfDeepFlavourTagInfos.clone(
     shallow_tag_infos = 'pfDeepCSVNegativeTagInfos',
