+++ RecoBTag/MXNet/plugins/BuildFile.xml
@@ -1,8 +1,7 @@
 <use name="FWCore/Framework"/>
 <use name="FWCore/MessageLogger"/>
-<library file="*.cc" name="RecoBTagTensorFlowPlugins">
+<library file="*.cc" name="RecoBTagDeepBoostedJetPlugins">
   <use name="DataFormats/BTauReco"/>
-  <use name="RecoBTag/TensorFlow"/>
-  <use name="PhysicsTools/TensorFlow"/>
+  <use name="PhysicsTools/MXNet"/>
   <flags EDM_PLUGIN="1"/>
 </library>
