+++ RecoBTag/MXNet/python/pfDeepBoostedJet_cff.py
@@ -1,10 +1,10 @@
 import FWCore.ParameterSet.Config as cms
 
-from RecoBTag.DeepBoostedJet.pfDeepBoostedJetTagInfos_cfi import pfDeepBoostedJetTagInfos
-from RecoBTag.DeepBoostedJet.pfDeepBoostedJetTags_cfi import pfDeepBoostedJetTags as _pfDeepBoostedJetTags
-from RecoBTag.DeepBoostedJet.pfDeepBoostedJetPreprocessParams_cfi import pfDeepBoostedJetPreprocessParams
-from RecoBTag.DeepBoostedJet.pfDeepBoostedDiscriminatorsJetTags_cfi import pfDeepBoostedDiscriminatorsJetTags
-from RecoBTag.DeepBoostedJet.pfMassDecorrelatedDeepBoostedDiscriminatorsJetTags_cfi import pfMassDecorrelatedDeepBoostedDiscriminatorsJetTags
+from RecoBTag.FeatureTools.pfDeepBoostedJetTagInfos_cfi import pfDeepBoostedJetTagInfos
+from RecoBTag.MXNet.pfDeepBoostedJetTags_cfi import pfDeepBoostedJetTags as _pfDeepBoostedJetTags
+from RecoBTag.MXNet.pfDeepBoostedJetPreprocessParams_cfi import pfDeepBoostedJetPreprocessParams
+from RecoBTag.MXNet.pfDeepBoostedDiscriminatorsJetTags_cfi import pfDeepBoostedDiscriminatorsJetTags
+from RecoBTag.MXNet.pfMassDecorrelatedDeepBoostedDiscriminatorsJetTags_cfi import pfMassDecorrelatedDeepBoostedDiscriminatorsJetTags
 
 # nominal DeepAK8
 pfDeepBoostedJetTags = _pfDeepBoostedJetTags.clone(
